# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WISMO API",
    "version": "1.0.0",
    "description": "Where Is My Order — AI-powered customer service for manufacturers. All endpoints require Firebase ID token authentication unless noted otherwise."
  },
  "servers": [
    { "url": "https://app.trywismo.co", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "firebaseToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Firebase ID token passed in request body as `idToken`"
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET for cron and webhook endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/quickbooks/invoice": {
      "get": {
        "summary": "Look up invoice by PO number",
        "tags": ["QuickBooks"],
        "parameters": [
          { "name": "po", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Invoice data or null" } }
      }
    },
...
  },
...
}